\name{rasterFW-package}

\alias{rasterFW}

\docType{package}

\title{ 
The rasterFW package
}

\description{
The rasterFW package links FWtools with the raster package

You need to have the FWtools installed for this to work. You can download these here: \url{http://fwtools.maptools.org/}

The functions provide simple wrappers around some of the command line programs included in FWtools. In many cases it may be easier to run the command line programs directly, or perhaps as an argument to a call to the R function \code{\link{system}}
}


\author{
Robert J. Hijmans \email{r.hijmans@gmail.com}
}

\keyword{package}
\keyword{spatial}
